Turned some size_t's back to ints, so that decrement loops could work properly
git-svn-id: svn://svn.code.sf.net/p/marathon/code/trunk@1531 a75d40c9-0969-4759-85e5-97cb871b7ed4

+++ Source_Files/ModelView/Dim3_Loader.cpp
@@ -238,8 +238,8 @@ bool LoadModel_Dim3(FileSpecifier& Spec, Model3D& Model, int WhichPass)
 		for (size_t ib=0; ib<NumBones; ib++)
 		{		
 			// Scan down the bone stack to find a bone that's the parent of some unlisted bone;
-			size_t ibsrch = NumBones;	// "Bone not found" value
-			size_t ibstck = -1;		// Empty stack
+			int ibsrch = NumBones;	// "Bone not found" value
+			int ibstck = -1;		// Empty stack
 			for (ibstck=StackTop; ibstck>=0; ibstck--)
 			{
 				// Note: the bone stack is indexed relative to the original,
